Refactor the textual-plotext full-plot logic to a resource_plot_pane widget
